Added Task Api Calls and configuration, moved modal to seperate file to reduce cluter, created constant file for ofline input
