+++ process.env
@@ -1,0 +1,1 @@
+REACT_APP_API_HOST=http://localhost:8088
