+++ src/apiService/constant.json
@@ -1,0 +1,3 @@
+{
+  "apiHost": "http://localhost:8088"
+}
