+++ src/components/task/task.js
@@ -1,7 +1,4 @@
-import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
-import Typography from "@mui/material/Typography";
-import Modal from "@mui/material/Modal";
 import React, { useEffect } from "react";
 import { useFormik } from "formik";
 import * as yup from "yup";
@@ -16,140 +13,45 @@ import {
   Select,
   TextField,
 } from "@mui/material";
-const style = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: 400,
-  bgcolor: "background.paper",
-  border: "2px solid #000",
-  boxShadow: 24,
-  p: 4,
-};
+import TaskModal from "./modal";
+import { GET } from "../../apiService/apiCall";
+import { apiServices } from "../../apiService/apiConfig";
 
 const Task = (props) => {
-  const [open, setOpen] = React.useState(false);
   const [user, setUser] = React.useState({});
   const [task, setTask] = React.useState({});
+  const [userTasks, setUserTasks] = React.useState([]);
 
-  const handleChangeTask = (event) => {
-    setTask(event.target.value);
-  };
+  const useTask = userTasks.map((task) => {
+    return (
+      <MenuItem key={task.id} value={task.id}>
+        {task.taskName}
+      </MenuItem>
+    );
+  });
 
-  const handleOpen = () => setOpen(true);
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  // Similar to componentDidMount and componentDidUpdate:
   useEffect(() => {
     var user_data = JSON.parse(localStorage.getItem("user_data"));
     if (user_data !== null) {
       setUser(user_data);
+      const getTasks = async () => {
+        let userTask = await GET(apiServices.Task.getAllTask);
+        console.log(userTask);
+      };
+      getTasks();
     } else {
-      console.log("no data");
-      handleOpen();
       setUser({});
     }
   }, []);
-  const handleFormSubmit = () => {
-    var data = {
-      name: formik.values.name,
-      email: formik.values.email,
-    };
-    localStorage.setItem("user_data", JSON.stringify(data));
-    handleClose();
+  const handleChangeTask = (event) => {
+    setTask(event.target.value);
   };
-  const validationSchema = yup.object({
-    email: yup
-      .string("Enter your email")
-      .email("Enter a valid email")
-      .required("Email is required"),
-  });
-  const formik = useFormik({
-    initialValues: {
-      email: "",
-      name: "",
-    },
-    validationSchema: validationSchema,
-    onSubmit: (values) => {
-      handleFormSubmit(values);
-    },
-  });
+
+  // Similar to componentDidMount and componentDidUpdate:
 
   return (
     <div>
-      <Modal
-        open={open}
-        onClose={handleClose}
-        aria-labelledby="modal-modal-title"
-        aria-describedby="modal-modal-description"
-      >
-        <Box sx={style}>
-          <Typography id="modal-modal-title" variant="h6" component="h6">
-            Enter Name and Email
-          </Typography>
-          <br></br>
-          <Box
-            sx={{
-              "& > :not(style)": { m: 2, height: "5ch" },
-            }}
-            id="modal-modal-description"
-          >
-            <Grid item xs={12}>
-              <TextField
-                fullWidth
-                size="small"
-                variant="outlined"
-                label="Email"
-                id="email"
-                name="email"
-                type="email"
-                inputProps={{ style: { textTransform: "lowercase" } }}
-                onChange={formik.handleChange}
-                onBlur={formik.handleBlur}
-                value={formik.values.email}
-                error={formik.touched.email && Boolean(formik.errors.email)}
-                helperText={formik.touched.email && formik.errors.email}
-                InputLabelProps={{
-                  shrink: true,
-                }}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                fullWidth
-                size="small"
-                variant="outlined"
-                label="Name"
-                id="name"
-                name="name"
-                type="text"
-                inputProps={{ style: { textTransform: "lowercase" } }}
-                onChange={formik.handleChange}
-                onBlur={formik.handleBlur}
-                value={formik.values.name}
-                error={formik.touched.name && Boolean(formik.errors.name)}
-                helperText={formik.touched.name && formik.errors.name}
-                InputLabelProps={{
-                  shrink: true,
-                }}
-              />
-            </Grid>
-          </Box>
-          <Divider style={{ marginBottom: "5px" }} />
-
-          <Button
-            onClick={formik.handleSubmit}
-            style={{ float: "right" }}
-            variant="contained"
-          >
-            Continue
-          </Button>
-        </Box>
-      </Modal>
-
+      <TaskModal />
       <Grid
         container
         spacing={0}
@@ -162,7 +64,7 @@ const Task = (props) => {
           <Card
             sx={{
               "& > :not(style)": {
-                minHeight: "500px",
+                minHeight: "550px",
                 m: 2,
                 width: "90ch",
                 height: "5ch",
@@ -194,6 +96,21 @@ const Task = (props) => {
                     labelId="demo-simple-select-label"
                     id="demo-simple-select"
                     value={task}
+                    label="Task"
+                    onChange={handleChangeTask}
+                  >
+                    {Boolean(userTasks.length) && useTask}
+                  </Select>
+                </FormControl>
+              </Grid>
+              <Grid sx={12}>
+                <FormControl fullWidth>
+                  <InputLabel id="demo-simple-select-label">
+                    Choose Language
+                  </InputLabel>
+                  <Select
+                    labelId="select-label-language"
+                    id="select-language"
                     label="Task"
                     onChange={handleChangeTask}
                   >
